backend/requirements.txt
# ─────────────────────────────────────────────────────
# Core Web Framework and API Utilities
# ─────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests>=2.32.3
aiohttp>=3.10.5
httpx>=0.27.2
websocket-client>=1.6.0
cryptography>=42.0.0
psutil>=5.9.0
AppOpener>=1.7.0
yfinance
pandas
openpyxl>=3.1.5
openai
python-docx
python-pptx
reportlab
PyPDF2
docx2pdf
